+++ crates/objc2/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 * Support `#[cfg(...)]` attributes in `extern_class!` macro.
+* Added support for selectors similar to `abc::def:` in the `sel!` macro.
 
 ### Changed
 * **BREAKING**: Using the automatic `NSError**`-to-`Result` functionality in